// src/app/models/identity.model.ts
export class Identity {
    id: number;
    titleP: string;
    emitter: string;
    place: string;
    valueT: string;
    value: string;
    addDateT: string;
    addDate: string;
    endDateT: string;
    endDate: string;
    level: string;
    iconsStars: Array<any>;
    credentialAssigned: boolean;
    isExpanded: boolean;
    isHidden: boolean;
    isChecked: boolean;
}